+++ requests/cart_item.go
@@ -12,6 +12,6 @@ type UpdateCartItem struct {
 	ProductId   uuid.UUID
 	ProductName string
 	ImageUrl    string
-	Price       int
-	Quantity    int
+	Price       float64
+	Quantity    int64
 }
